Rework CORS behaviour for getLatest
We previously tried to do a single HEAD request alone to the latest
endpoint so we can parse out the build number. This is incompatible with
CORS, since redirects should be transparent and when setting redirects
to manual we always get an opaqueredirect response, which is
indistinguishable from a network error.

The solution here is to go full CORS and follow the redirect
transparently, ending with a HEAD request to the final url. From the
fetch response we can the access the url attribute and parse the build
number from there. This takes two requests unfortunately since the
browser follows the redirect.